+++ extension/typedef.js
@@ -1,45 +1,74 @@
 /**
  * @typedef {Object} SDKContext
- * @property {Object} config
- * @property {string} config.shopifyShopAlias
- * @property {string} config.shopifyAccessToken
- * @property {string} config.userRegistrationUrl
- * @property {string} config.userDataCacheTtl
- * @property {string} config.stage
- * @property {string} config.credentials.baseDomain
- * @property {string} config.credentials.clientId
- * @property {string} config.credentials.clientSecret
- * @property {string} config.credentials.refreshToken
- * @property {Object} config.requestTimeout
- * @property {number} config.requestTimeout.token
- * @property {number} config.requestTimeout.bigApi
- * @property {Object} meta
- * @property {string|number} meta.userId
- * @property {string} meta.appId
- * @property {Object} storage
- * @property {Object} storage.user
- * @property {Function} storage.user.get - (string key)
- * @property {Function} storage.user.set - (string key, mixed value)
- * @property {Function} storage.user.del - (string key)
- * @property {Object} storage.extension
- * @property {Function} storage.extension.get - (string key)
- * @property {Function} storage.extension.set - (string key, mixed value)
- * @property {Function} storage.extension.del - (string key)
- * @property {Function} storage.extension.map.get - (string mapName)
- * @property {Function} storage.extension.map.set - (string mapName, Object map)
- * @property {Function} storage.extension.map.del - (string mapName)
- * @property {Function} storage.extension.map.getItem - (string mapName, string key)
- * @property {Function} storage.extension.map.setItem - (string mapName, string key, string value)
- * @property {Function} storage.extension.map.delItem - (string mapName, string key)
- * @property {Object} log
- * @property {Function} log.trace
- * @property {Function} log.debug
- * @property {Function} log.info
- * @property {Function} log.warn
- * @property {Function} log.error
- * @property {Function} log.fatal
+ * @property {ExtensionConfig} config
+ * @property {SDKContextMeta} meta
+ * @property {SDKContextStorage} storage
+ * @property {SDKContextLog} log
+ * @property {Function} tracedRequest
  */
 
+/**
+ * @typedef {Object} ExtensionConfig
+ * @property {string} shopifyShopAlias
+ * @property {string} shopifyAccessToken
+ * @property {string} userRegistrationUrl
+ * @property {string} userDataCacheTtl
+ * @property {string} stage
+ * @property {Object} credentials
+ * @property {string} credentials.baseDomain
+ * @property {string} credentials.clientId
+ * @property {string} credentials.clientSecret
+ * @property {string} credentials.refreshToken
+ * @property {Object} requestTimeout
+ * @property {number} requestTimeout.token
+ * @property {number} requestTimeout.bigApi
+ */
+
+/**
+ * @typedef {Object} SDKContextMeta
+ * @property {string} deviceId
+ * @property {string} appId
+ * @property {string} userId
+ * @property {string} appLanguage
+ */
+
+/**
+ * @typedef {Object} SDKContextStorage
+ * @property {SDKContextEntityStorage} extension
+ * @property {SDKContextEntityStorage} device
+ * @property {SDKContextEntityStorage} user
+ */
+
+/**
+ * @typedef {Object} SDKContextEntityStorage
+ * @property {Function} get - (string key, Function cb)
+ * @property {Function} set - (string key, mixed value, Function cb)
+ * @property {Function} del - (string key, Function cb)
+ * @property {Object} map
+ * @property {Function} map.get - (string mapName)
+ * @property {Function} map.set - (string mapName, Object map)
+ * @property {Function} map.del - (string mapName)
+ * @property {Function} map.getItem - (string mapName, string key)
+ * @property {Function} map.setItem - (string mapName, string key, string value)
+ * @property {Function} map.delItem - (string mapName, string key)
+ */
+
+/**
+ * @typedef {Object} SDKContextLog
+ * @property {Function} trace
+ * @property {Function} debug
+ * @property {Function} info
+ * @property {Function} warn
+ * @property {Function} error
+ * @property {Function} fatal
+ */
+
+/**
+ * @typedef {Object} SgxsMeta
+ * @property {string} sessionId
+ */
+
+// ========= Shopify Specifics ==========
 /**
  * @typedef {Object} ShopifyAddress
  * @property {Object} address
@@ -61,6 +90,22 @@
  */
 
 /**
+ * @typedef {Object} ShopifyCustomer
+ * @property {string} id The user ID in the format "gid://shopify/Customer/<id>" and base64 encoded
+ * @property {string} email
+ * @property {string} firstName
+ * @property {string} lastName
+ * @property {string|null} phone
+ */
+
+/**
+ * @typedef {Object} ShopifyCustomerAccessToken
+ * @property {string} accessToken
+ * @property {string} expiresAt
+ */
+
+// ========= Shopgate Specifics ==========
+/**
  * @typedef {Object} ShopgateAddress
  * @property {Object} address
  * @property {number} [id]
@@ -79,22 +124,20 @@
  */
 
 /**
+ * @typedef {Object} ShopgateUser
+ * @property {string} firstName
+ * @property {string} lastName
+ * @property {ShopgateUserCustomAttributes} customAttributes
+ */
+
+/**
  * @typedef {Object} ShopgateAddressCustomAttributes
  * @property {string} [company]
  * @property {string} [phone]
  */
 
 /**
- * @typedef {Object} ShopifyCustomer
- * @property {string} id The user ID in the format "gid://shopify/Customer/<id>" and base64 encoded
- * @property {string} email
- * @property {string} firstName
- * @property {string} lastName
- * @property {string|null} phone
- */
-
-/**
- * @typedef {Object} ShopifyCustomerAccessToken
- * @property {string} accessToken
- * @property {string} expiresAt
+ * @typedef {Object} ShopgateUserCustomAttributes
+ * @property {string} [email]
+ * @property {string} [phone]
  */