+++ extension/user/address/add.js
@@ -1,7 +1,7 @@
-const Tools = require('../lib/tools')
-const UnauthorizedError = require('../models/Errors/UnauthorizedError')
-const ApiFactory = require('../lib/shopify.api.factory')
-const { mapCountry, mapCustomAttributes } = require('../lib/mapper')
+const Tools = require('../../lib/tools')
+const UnauthorizedError = require('../../models/Errors/UnauthorizedError')
+const ApiFactory = require('../../lib/shopify.api.factory')
+const { mapCountry, mapCustomAttributes } = require('../../lib/mapper')
 
 /**
  * @param {SDKContext} context